umpf_pipeline/requirements.txt
python-dotenv
openai>=1.0.0
pdfminer.six
rich
